+++ packages/api/tests/endpoints/test-granules.js
@@ -1861,7 +1861,7 @@ test.serial('create (POST) rejects if a granule already exists in postgres', asy
   const errorText = JSON.parse(response.error.text);
   t.is(errorText.statusCode, 409);
   t.is(errorText.error, 'Conflict');
-  t.is(errorText.message, `A granule already exists for granule_id: ${newGranule.granuleId} with collection ID ${t.context.collectionId}`);
+  t.is(errorText.message, `A granule already exists for granuleId: ${newGranule.granuleId}`);
 });
 
 test.serial('create (POST) returns bad request if a granule is submitted with a bad collectionId', async (t) => {
@@ -1881,7 +1881,7 @@ test.serial('create (POST) returns bad request if a granule is submitted with a 
   t.is(response.error.message, 'cannot POST /granules (400)');
 });
 
-test.serial('create (POST) rejects if a granule with same granuleId already exists in postgres', async (t) => {
+test.serial('create (POST) rejects if a granule with same granuleId but different collectionId already exists in postgres', async (t) => {
   const { collectionPgModel } = t.context;
   const newGranule = fakeGranuleFactoryV2({
     collectionId: t.context.collectionId,
@@ -1934,7 +1934,7 @@ test.serial('create (POST) rejects if a granule with same granuleId already exis
   const errorText = JSON.parse(response.error.text);
   t.is(errorText.statusCode, 409);
   t.is(errorText.error, 'Conflict');
-  t.is(errorText.message, `A granule already exists for granule_id: ${newGranule.granuleId} across another collection.`);
+  t.is(errorText.message, `A granule already exists for granuleId: ${newGranule.granuleId}`);
 });
 
 test.serial('PUT replaces an existing granule in all data stores', async (t) => {