+++ packages/api/tests/lib/writeRecords/test-write-granules.js
@@ -1855,7 +1855,7 @@ test.serial('writeGranulesFromMessage does not write a granule to Postgres or Dy
     granulePgModel,
   }));
 
-  t.true(error.message.includes(`A granule already exists for granule_id: ${pgGranule.granule_id} across another collection`));
+  t.true(error.message.includes(`A granule already exists for granuleId: ${pgGranule.granule_id} with collectionId ${collectionCumulusId}`));
   t.false(await granuleModel.exists({ granuleId }));
   t.false(
     await t.context.granulePgModel.exists(knex, {