+++ CHANGELOG.md
@@ -44,8 +44,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - **CUMULUS-2944**
   - Added configuration to increase the limit for body-parser's JSON and URL encoded parsers to allow for larger input payloads
 -**CUMULUS-3100**
-  - Updated `POST` granules endpoint to check if granuleId exists across a different collection.
-  - Udpated logic for writing granules from a message to check if granuleId exists across a different collection.
+  - Updated `POST` granules endpoint to check if granuleId exists across all collections rather than a single collection.
+  - Udpated logic for writing granules from a message to check if granuleId exists across a different collection and throw conflict error if so.
 
 ### Changed